+++ governance/root_manager.go
@@ -90,13 +90,18 @@ type RootManager struct {
 	rootQuotaLock           sync.Mutex
 	ProposalQuotaTimeWindow time.Duration
 	exclusionQuotaLock      sync.Mutex
+
+	ApprovalMaxFailures    uint64
+	transitionBlockChecker *transitionBlockChecker
 }
 
 // Config of root manager
 type Config struct {
-	RootList                common.RootList `toml:"-"`
-	ProposalQuotaMax        uint
-	ProposalQuotaTimeWindow time.Duration
+	RootList                      common.RootList `toml:"-"`
+	ProposalQuotaMax              uint
+	ProposalQuotaTimeWindow       time.Duration
+	ApprovalMaxFailures           uint64
+	TransitionBlockVerifiedBlocks uint64
 }
 
 type DiffEntry struct {
@@ -169,7 +174,14 @@ func NewRootManager(am *accounts.Manager, networkId uint64, datadir string, cfg 
 
 		ProposalQuotaMax:        cfg.ProposalQuotaMax,
 		ProposalQuotaTimeWindow: cfg.ProposalQuotaTimeWindow,
+
+		ApprovalMaxFailures: cfg.ApprovalMaxFailures,
 	}
+
+	manager.transitionBlockChecker = newTransitionBlockChecker(
+		int64(cfg.TransitionBlockVerifiedBlocks),
+		manager.HandleTransitionBlockSignature,
+	)
 
 	manager.setMaxRewindLimit(uint64(params.FullImmutabilityThreshold))
 
@@ -184,6 +196,7 @@ func (s *RootManager) InitDownloader(dl *downloader.Downloader) {
 
 func (s *RootManager) InitBlockChain(bc *core.BlockChain) {
 	s.bc = bc
+	s.transitionBlockChecker.initBlockChain(bc)
 }
 
 func (s *RootManager) InitRegistry(reg *contracts.Registry) {
@@ -1163,6 +1176,15 @@ func (s *RootManager) ValidatePreviousTransitionBlockSignature() {
 }
 
 func (s *RootManager) HandleTransitionBlockSignature(header *types.Header) {
+	checkLater, err := s.transitionBlockChecker.checkTransitionBlockLater(header)
+	if err != nil {
+		log.Error("Failed to check transition block", "err", err)
+		return
+	}
+	if checkLater {
+		return
+	}
+
 	if header == nil {
 		log.Debug("Nil header is passed to HandleTransitionBlockSignature")
 		return